Added update button to web interface.
Added a mapping to the gplaycli function which updates all the downloaded apks. Set the cli.yes flag to avoid timeouts due to missing command line confirmation. One thing to improve is that when all the apps are up to date, the web app returns ERR_EMPTY_RESPONSE, while it would be nice to have a message printed on the webpage.